more sane testing setup
