+++ pyproject.toml
@@ -29,6 +29,7 @@ pytest = "^5.1"
 mdv = "^1.7.4"
 rich = "^1.0.0"
 pytest-cov = "^2.9.0"
+pytest-asyncio = "^0.12.0"
 
 [build-system]
 requires = ["poetry>=0.12"]
